% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_version}
\alias{find_version}
\title{Find the Past Version That Is Closest to a Given Date}
\usage{
find_version(ref_date, tags, return_as_tag = FALSE)
}
\description{
Find the Past Version That Is Closest to a Given Date
}
