% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iso_3166_alpha_2_countries.R
\name{get_iso_3166_alpha_2_countries}
\alias{get_iso_3166_alpha_2_countries}
\title{Get ISO 2-Digit Country codes}
\usage{
get_iso_3166_alpha_2_countries()
}
\description{
Get ISO 2-Digit Country codes
}
