% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_remote_archives.R
\name{list_remote_archives}
\alias{list_remote_archives}
\title{List All OPENTSI Compliant Archives of an Organization}
\usage{
list_remote_archives(org = "opentsi", max_n = 100, git_platform = "github")
}
\arguments{
\item{org}{character name of an organization}

\item{max_n}{numeric max number of archives to be listed}

\item{git_platform}{character can only be set to GitHub for now.}
}
\description{
List All OPENTSI Compliant Archives of an Organization
}
