% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_time_series.R
\name{list_time_series}
\alias{list_time_series}
\title{List Keys and Minimal Descriptions of all Time Series in an Archive}
\usage{
list_time_series(repo, remote_org = "opentsi")
}
\arguments{
\item{repo}{character archive repo name}

\item{remote_org}{character name of the GitHub Org that hosts the time series.
Set to NULL to interpret repo as a local path a repository.}
}
\description{
List Keys and Minimal Descriptions of all Time Series in an Archive
}
