% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{print_resolved}
\alias{print_resolved}
\title{Print Resolved Metadata}
\usage{
print_resolved(resolved, format = c("full", "short"))
}
\arguments{
\item{resolved}{Output from resolve_meta().}

\item{format}{Either "full" or "short".}
}
\value{
Invisibly returns the input.
}
\description{
Prints key-label mappings in a readable format.
}
