% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_history.R
\name{read_history}
\alias{read_history}
\title{Read Entire History of a Time Series}
\usage{
read_history(
  series,
  remote_archive = "opentsi/demo",
  rbind_dt = TRUE,
  add_suffix = TRUE,
  wide = TRUE,
  lastn = NULL
)
}
\description{
Alias for the read_open_ts function with date parameter set to NULL.
}
