% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_open_ts.R
\name{read_open_ts}
\alias{read_open_ts}
\title{Read Time Series from an Open Time Series Archive}
\usage{
read_open_ts(ts_keys, date_time = Sys.time(), repo, remote_org = "opentsi")
}
\arguments{
\item{series}{character vector containing the queried time series keys.}

\item{date}{date used to determine the version of a set of time series.
Defaults to Sys.Date().}

\item{remote_archive}{character contains the username/repo of a GitHub archive. Defaults to rseed-koflab/ch.kof.}

\item{rbind_dt}{boolean should data.tables be row bound to a single data.table?
Defaults to TRUE.}

\item{add_suffix}{boolean should a version suffix be added to the id? Defaults to FALSE.}
}
\description{
This function reads time series from an Open Time Series Archive. It is
a convenient way to read directly from a remote Git repository or from local
open time series archive. As a convenience function it is uses reasonable defaults
using the official archives. Hence calls need to be adapted if you want to use your
archives.
}
