% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tags_list}
\alias{tags_list}
\title{Get a List of All Tags from a Remote Time Series Archive}
\usage{
tags_list(remote_archive)
}
\description{
For now, opentimeseries only supports GitHub as an arvchive host.
We're working to suport other hosts including self-hosted git platforms.
}
