% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_commit_dates}
\alias{get_commit_dates}
\title{Get Commit Hashes and Dates from Remote Repository}
\usage{
get_commit_dates(remote_archive = "opentsi/kofethz", lastn = 100)
}
\description{
Get Commit Hashes and Dates from Remote Repository
}
